refactor: rename 'Generics' to 'Generic Types'
