+++ README.adoc
@@ -20,7 +20,7 @@ This project is for the course "Java Basics". It contains for each theme of this
 |7. OOP: Classes and Objects.|link:src/main/resources/oop-classes-and-objects.adoc[oop-classes-and-objects.adoc]
 |8. Strings.|link:src/main/resources/strings.adoc[strings.adoc], link:src/main/java/com/rakovets/course/javabasics/practice/strings[package: strings]
 |9. OOP: Inheritance and Polymorphism.|link:src/main/resources/oop-inheritance-and-polymorphism.adoc[oop-inheritance-and-polymorphism.adoc]
-|10. Generic Types (link:https://jcp.org/en/jsr/detail?id=14[JSR 14]).|link:src/main/resources/generics.adoc[generics.adoc]
+|10. Generic Types (link:https://jcp.org/en/jsr/detail?id=14[JSR 14]).|link:src/main/resources/generic-types.adoc[generic-types.adoc]
 |11. Java Collection Framework: Collection.|link:src/main/resources/jcf-collection.adoc[jcf-collection.adoc]
 |12. Java Collection Framework: List.|link:src/main/resources/jcf-list.adoc[jcf-list.adoc]
 |13. Java Collection Framework: Set.|link:src/main/resources/jcf-set.adoc[jcf-set.adoc]
