+++ src/main/java/com/rakovets/course/javabasics/example/generictypes/model/Transaction.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.javabasics.example.generics.model;
+package com.rakovets.course.javabasics.example.generictypes.model;
 
 public class Transaction {
     public static boolean executeTransaction(BankAccount from, BankAccount to, int amount) {
