+++ src/main/resources/generic-types.adoc
